Synchronize changes with mvp-v1-staging (#34)
* chore: Monorepo configuration files and documentation

* feat: Backend

user-management, database, auth, data-collectors, interviews, glossaries and document-analysis modules

* fix: Backend

Update some files to deploy the API (NestJS)

* feat: Frontend

The frontend is not finished. Middlewares do not work.

* fix: Frontend

Fix some bugs and improve some parts. This frontend is not finished. Middlewares do not work and reloading causes the user state to be lost.

* fix: Backend

This backend is not finished

* fix: Backend

Fix some bugs to deploy in a staging environment. This backend is not finished.

* fix: Frontend

Fix some bugs and improve some parts to deploy the app in a staging environment. This frontend is not fished. Middlewares do not work.

* chore: Remove some .example files

.prod.env.example and .stag.env.example

* fix: Backend

Change --production flag by --omit=dev in RUN npm install

* fix: Backend

Change RUN command to try deploying in a staging environment again

* fix: Backend

Update some dependencies to fix a deploy error

* fix: Backend

Install @nestjs/cli before the rest of the dependencies to fix a deploy error

* fix: Backend

Change CMD command and add ENV to load an environment variable

* fix: Frontend

Add Dockerfile and update some dependencies

* fix: Frontend

Remove --omit=dev of RUN npm install command to fix a deploy error

* fix: Frontend

Comment some lines to fix a deploy error

* fix: Frontend

Comment some lines in /(auth)/profiles/page.tsx to fix a deploy error

* fix: Frontend

Add Promise to params in ModuleProps to fix a deploy error

* fix: Frontend

Fix type error

* fix: Frontend

Change type of MouseEvent to HTMLDivElement to fix a deploy error

* fix: Frontend

Create a LoginClient component and add Suspense to /(auth)/login/page.tsx, because useSearchParams forces a render on the client and when used on a page Next.js tries to pre-render on the server and this causes an error

* fix: Frontend

Add 2 lines in the Dockerfile to define the environment variables for the build to fix an error in the deployed app

* fix: Frontend

Fix some errors (infinite loop) and question types implementation beginning

* fix: Backend

Fix some errors and question types implementation beginning

* fix: Frontend

Continue fixing and improving different things. The frontend is not finished. Middlewares do not work and there may be some errors.

* fix: Backend

Continue fixing and improving different things. The backend is not finished. There may be some errors.

* fix: Frontend

Continue fixing and improving different things such as team management system for 'Individual' user (there are still some errors). The frontend is not finished. Middlewares do not work.

* fix: Backend

Continue fixing and improving different things, for example, all files that are neccesary to the team management system for 'Individual' users (there are still some errors)

* fix: Frontend

Continue fixing and improving different things (there are still errors). The frontend is not finished. Middlewares do not work.

* fix: Backend

Continue fixing and improving different things (there are still errors). The backend is not finished.

* fix: Frontend

Continue fixing and improving different things such as /(companies) pages and questions creation. The frontend is not finished. Middlewares do not work.

* fix: Backend

Continue fixing and improving different things such as questions creation. The backend is not finished, there are still errors.

* fix: Frontend

Continue fixing and implementing different things such as team and company management or forms and surveys creation. The frontend is not finished. Middlewares do not work.

* fix: Backend

Continue fixing and improving different things such as questions creation and deletion. The backend is not finished, there may be some errors.

* fix: Frontend

Fixing error with  /register/page.tsx

* fix: Frontend

Continue improving and fixing different things. The frontend is not finished. Middlewares not found.

* fix: Backend

Continue improving and fixing different things. The backend is not finished, there may be some errors.